brain_server/pyproject.toml
[project]
name = "brain_server"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "alibabacloud-ecs20140526 (>=6.1.0,<7.0.0)",
    "docker (>=7.1.0,<8.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "pymoo (>=0.6.1.3,<0.7.0.0)",
    "pyarrow (>=19.0.1,<20.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "duckdb (>=1.2.2,<2.0.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "grpcio (>=1.71.0,<2.0.0)",
    "grpcio-tools (>=1.71.0,<2.0.0)",
    "cython (>=3.0.12,<4.0.0)",
]

[tool.poetry]
packages = [{include = "brain_server", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
